restructure LLM type module packaging to be easier to extend and build on top of - rk
